% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Events.R, R/Outcomes.R,
%   R/Predictions.R, R/Sensors.R
\name{columns}
\alias{columns}
\alias{deployments}
\alias{deployments,Events-method}
\alias{deployments,Outcomes-method}
\alias{deployments,Predictions-method}
\alias{deployments,Sensors-method}
\alias{columns,Sensors-method}
\title{Accessors}
\usage{
columns(object)

deployments(object)

\S4method{deployments}{Events}(object)

\S4method{deployments}{Outcomes}(object)

\S4method{deployments}{Predictions}(object)

\S4method{deployments}{Sensors}(object)

\S4method{columns}{Sensors}(object)
}
\arguments{
\item{object}{The object to access}
}
\value{
\subsection{\code{columns}}{

A list of columns of bio-logging sensor data (e.g. pitch or overall dynamic
body acceleration) in \code{object}. Applicable to \code{Sensors} only.
}

\subsection{\code{deployments}}{

A list of bio-logger deployments in \code{objects}.
}
}
\description{
Access deployment and other information from \code{Events},
\code{Outcomes}, \code{Predictions}, and \code{Sensors} objects.
}
\examples{
c(lunge_sensors, lunge_events) \%<-\% load_lunges()
columns(lunge_sensors)
deployments(lunge_events)

}
