% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{as.data.frame.Predictions}
\alias{as.data.frame.Predictions}
\title{Convert Predictions to data.frame}
\usage{
\method{as.data.frame}{Predictions}(x, ...)
}
\arguments{
\item{x}{Predictions}

\item{...}{not used}
}
\value{
data.frame
}
\description{
Convert Predictions to data.frame
}
