% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensors.R
\name{as.data.frame.Sensors}
\alias{as.data.frame.Sensors}
\title{Convert Sensors to data.frame}
\usage{
\method{as.data.frame}{Sensors}(x, ...)
}
\arguments{
\item{x}{\link{Sensors}}
}
\value{
\link{data.frame}
}
\description{
Convert Sensors to data.frame
}
