% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Events.R, R/Sensors.R
\name{divide}
\alias{divide}
\alias{divide,Events-method}
\alias{divide,Sensors-method}
\title{Divide Events/Sensors objects by deployment IDs}
\usage{
divide(object, deployids)

\S4method{divide}{Events}(object, deployids)

\S4method{divide}{Sensors}(object, deployids)
}
\arguments{
\item{object}{\verb{[Events/Sensors]} An \code{\link{Events}} or
\code{\link{Sensors}} object.}

\item{deployids}{\verb{[character]} The deployment IDs to divide on. Each ID must
be a deployment in \code{object}. Deployments with IDs in \code{deployids} will end
up in one \code{Events/Sensors} object, the remaining deployments will be in
another.}
}
\value{
\verb{[list(Events/Sensors)]} A list of two \code{Events}/\code{Sensors} objects.
The first list element contains the deployments with IDs in \code{deployids},
the second element contains the remainder.
}
\description{
Useful for dividing bio-logging sensor data (\code{Sensors}) and behavioral event
data (\code{Events}) into test-train sets.
}
\examples{
c(lunge_sensors, lunge_events) \%<-\% load_lunges()
divided_sensors <- divide(lunge_sensors, "bw180905-53")
# A Sensors object with *only* deployment "bw180905-53"
deployments(divided_sensors[[1]])
# A Sensors object with the remaining five deployments
deployments(divided_sensors[[2]])

}
