% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R
\docType{class}
\name{Events-class}
\alias{Events-class}
\alias{Events}
\title{Events class}
\usage{
Events(event_data, deployid_col, datetime_col)
}
\arguments{
\item{event_data}{\verb{[data.frame]}. A two-column data frame with bio-logger
deployment IDs in one column and event datetimes in the other. The event
datetimes must be \code{POSIXct}.}

\item{deployid_col}{\verb{[character(1)]}. The name of the column in \code{event_data}
containing deployment IDs.}

\item{datetime_col}{\verb{[character(1)]}. The name of the column in \code{event_data}
with event datetimes.}
}
\description{
Class representation for labeled behavioral events in bio-logging data.
}
\details{
Create an Events object with \code{Events}().
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of Pandas DatetimeIndexes.}
}}

\examples{
# Pretend we have behavioral event data for two dogs that like to bark at a
# clock tower
barking_df <- data.frame(
  dogs = rep(c("A", "B", each = 4)),
  barktime = as.POSIXct("2000-01-01 06:00") + lubridate::hours(0:7)
)
barking_df
Events(barking_df, "dogs", "barktime")

}
