% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outcomes.R
\docType{class}
\name{Outcomes-class}
\alias{Outcomes-class}
\title{Outcomes class}
\description{
Class representation for prediction outcomes.
}
\details{
Users should not create Outcomes objects directly. They're the return type
of sb_assess().
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of Series.}
}}

