% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\docType{class}
\name{Predictions-class}
\alias{Predictions-class}
\alias{deployments,Predictions-method}
\title{Predictions class}
\usage{
\S4method{deployments}{Predictions}(object)
}
\description{
Class representation for event predictions.
}
\details{
Users should not create Predictions objects directly. They're the return type
of sb_predict().
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of tuples (Series, DatetimeIndex).}
}}

