% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stickleback.R
\name{sb_assess}
\alias{sb_assess}
\title{Assess Stickleback predictions}
\usage{
sb_assess(sb, predicted, events)
}
\arguments{
\item{sb}{\link{Stickleback}}

\item{predicted}{\link{Predictions}}

\item{events}{\link{Events}}
}
\value{
\link{Outcomes}
}
\description{
Assess Stickleback predictions
}
