% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sb_plot_data}
\alias{sb_plot_data}
\title{Plot Stickleback training data}
\usage{
sb_plot_data(deployid, sensors, events)
}
\arguments{
\item{deployid}{\verb{[character(1)]} Deployment ID.}

\item{sensors}{\verb{[Sensors]} Sensor data.}

\item{events}{\verb{[Events]} Event data.}
}
\value{
A plotly figure.
}
\description{
Creates an interactive plotly figure for examining sensor and event data.
}
