% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stickleback.R
\name{sb_predict}
\alias{sb_predict}
\title{Predict with a Stickleback model}
\usage{
sb_predict(sb, sensors)
}
\arguments{
\item{sb}{\link{Stickleback}}

\item{sensors}{\link{Sensors}}
}
\description{
Predict with a Stickleback model
}
