% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensors.R
\docType{class}
\name{Sensors-class}
\alias{Sensors-class}
\alias{Sensors}
\alias{deployments,Sensors-method}
\alias{columns,Sensors-method}
\title{Sensors class}
\usage{
Sensors(sensor_data, deployid_col, datetime_col, sensor_cols)

\S4method{deployments}{Sensors}(object)

\S4method{columns}{Sensors}(object)
}
\arguments{
\item{sensor_data}{a data frame. Must have columns for deployment ID and
datetime (POSIXct).}

\item{deployid_col}{character(1). Name of column with deployment ID.}

\item{datetime_col}{character(1). Name of column with event datetimes.}

\item{sensor_cols}{character. Names of columns containing sensor data.}
}
\description{
Class representation for bio-logging sensor data.
}
\details{
Create a Sensors object with Sensors().
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of Pandas DataFrames.}
}}

