% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R, R/Outcomes.R, R/Predictions.R,
%   R/Sensors.R
\name{show,Events-method}
\alias{show,Events-method}
\alias{show,Outcomes-method}
\alias{show,Predictions-method}
\alias{show,Sensors-method}
\title{Print an object}
\usage{
\S4method{show}{Events}(object)

\S4method{show}{Outcomes}(object)

\S4method{show}{Predictions}(object)

\S4method{show}{Sensors}(object)
}
\arguments{
\item{object}{Events, Outcomes, Predictions, or Sensors}
}
\description{
Print an object
}
