% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R
\name{split,Events-method}
\alias{split,Events-method}
\title{Split events by deployment IDs}
\usage{
\S4method{split}{Events}(object, ids)
}
\arguments{
\item{object}{\link{Events}}

\item{ids}{Deployment IDs in one split}
}
\value{
a list of two Events objects. The first element contains the
deployments with IDs in ids, the second element contains the remainder.
}
\description{
Useful for test-train splitting.
}
