% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensors.R
\name{split,Sensors-method}
\alias{split,Sensors-method}
\title{Split sensors by deployment IDs}
\usage{
\S4method{split}{Sensors}(object, ids)
}
\arguments{
\item{object}{\link{Sensors}}

\item{ids}{Deployment IDs in one split}
}
\value{
a list of two Sensors objects. The first element contains the
deployments with IDs in ids, the second element contains the remainder.
}
\description{
Useful for test-train splitting.
}
