% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R, R/Sensors.R
\name{split.Events}
\alias{split.Events}
\alias{split.Sensors}
\title{Split objects by deployment IDs}
\usage{
\method{split}{Events}(object, deployids)

\method{split}{Sensors}(object, deployids)
}
\arguments{
\item{object}{\link{Sensors/Events}}

\item{deployids}{Deployment IDs in one split}
}
\value{
a list of two Sensors/Events objects. The first element contains the
deployments with IDs in deployids, the second element contains the
remainder.
}
\description{
Useful for test-train splitting.
}
