% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Events.R, R/Outcomes.R,
%   R/Predictions.R, R/Sensors.R
\name{columns}
\alias{columns}
\alias{deployments}
\alias{deployments,Events-method}
\alias{deployments,Outcomes-method}
\alias{deployments,Predictions-method}
\alias{deployments,Sensors-method}
\alias{columns,Sensors-method}
\title{Accessors}
\usage{
columns(object)

deployments(object)

\S4method{deployments}{Events}(object)

\S4method{deployments}{Outcomes}(object)

\S4method{deployments}{Predictions}(object)

\S4method{deployments}{Sensors}(object)

\S4method{columns}{Sensors}(object)
}
\arguments{
\item{object}{The object to access}
}
\value{
\itemize{ \item \code{columns()}: a list of columns in sensor data
(Sensors only). \item \code{deployments()}: a list of deployment IDs. }
}
\description{
Access information from Events, Outcomes, Predictions, and
Sensors objects.
}
