% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R
\name{as.data.frame.Events}
\alias{as.data.frame.Events}
\title{Convert Events to data.frame}
\usage{
\method{as.data.frame}{Events}(x, ...)
}
\arguments{
\item{x}{Events}

\item{...}{not used}
}
\value{
data.frame
}
\description{
Convert Events to data.frame
}
