% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outcomes.R
\name{as.data.frame.Outcomes}
\alias{as.data.frame.Outcomes}
\title{Convert Outcomes to data.frame}
\usage{
\method{as.data.frame}{Outcomes}(x, ...)
}
\arguments{
\item{x}{Outcomes}

\item{...}{not used}
}
\value{
data.frame
}
\description{
Convert Outcomes to data.frame
}
