% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc.R
\name{compose_tsc}
\alias{compose_tsc}
\title{Compose time series classification model across multiple variables.}
\usage{
compose_tsc(module, algorithm, params, columns)
}
\arguments{
\item{module}{\link{character(1)} Name of module in sktime.classification, e.g.
"interval_based".}

\item{algorithm}{\link{character(1)} Name of time series classification algorithm.}

\item{params}{\link{list} Hyperparameters to algorithm.}

\item{columns}{\link{character} Names of columns for composition.}
}
\value{
A time series classification model, for use in fitting a Stickleback
object.
}
\description{
Compose time series classification model across multiple variables.
}
