% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc.R
\name{create_tsc}
\alias{create_tsc}
\title{Create a time series classification model.}
\usage{
create_tsc(module, algorithm, params)
}
\arguments{
\item{module}{\link{character(1)} Name of module in sktime.classification, e.g.
"interval_based".}

\item{algorithm}{\link{character(1)} Name of time series classification algorithm.}

\item{params}{\link{list} Hyperparameters to algorithm.}
}
\value{
A time series classification model, for use in fitting a Stickleback
object.
}
\description{
Create a time series classification model.
}
