% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/Events.R, R/Sensors.R
\name{divide}
\alias{divide}
\alias{divide,Events-method}
\alias{divide,Sensors-method}
\title{Divide Events/Sensors objects by deployment IDs}
\usage{
divide(object, deployids)

\S4method{divide}{Events}(object, deployids)

\S4method{divide}{Sensors}(object, deployids)
}
\arguments{
\item{object}{Events/Sensors}

\item{deployids}{Vector of deployment IDs}
}
\value{
a list of two Events/Sensors objects. The first element contains the
deployments with IDs in deployids, the second element contains the
remainder.
}
\description{
Useful for dividing into test-train sets.
}
