% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R
\docType{class}
\name{Events-class}
\alias{Events-class}
\alias{Events}
\title{Events class}
\usage{
Events(event_data, deployid_col, datetime_col)
}
\arguments{
\item{event_data}{a data frame. Must have columns for deployment ID and event
datetimes (POSIXct).}

\item{deployid_col}{character(1). Name of column with deployment ID.}

\item{datetime_col}{character(1). Name of column with event datetimes.}
}
\description{
Class representation for labeled behavioral events in bio-logging data.
}
\details{
Create an Events object with Events().
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of Pandas DatetimeIndexes.}
}}

