% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_lunges}
\alias{load_lunges}
\title{Load sample data: blue whale lunge-feeding}
\usage{
load_lunges()
}
\value{
A list with elements \code{sensors} (list of data frames containing sensor
data, named by deployment ID) and \code{events} (list of labeled events, also
named by deployment ID).
}
\description{
Load sample data: blue whale lunge-feeding
}
\examples{
sample_data <- load_lunges()
}
