% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outcomes.R
\docType{class}
\name{Outcomes-class}
\alias{Outcomes-class}
\title{Outcomes class}
\description{
Class representation for prediction outcomes. \code{Outcomes} objects are created
by \code{\link{sb_assess}}. They hold information about the outcomes (true
positive, false positive, false negative) of Stickleback predictions. You
shouldn't create \code{Outcomes} objects directly.
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of Pandas Series.}
}}

\seealso{
\code{\link{sb_assess}}
}
