% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\docType{class}
\name{Predictions-class}
\alias{Predictions-class}
\title{Predictions class}
\description{
Class representation for Stickleback predictions. \code{Predictions} objects are
created by \code{\link{sb_predict}}. You shouldn't create \code{Predictions}
objects directly.
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of tuples (Pandas Series, Pandas DatetimeIndex).}
}}

\seealso{
\code{\link{sb_predict}}, \code{\link{sb_plot_predictions}}
}
