% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stickleback.R
\name{sb_fit}
\alias{sb_fit}
\title{Fit a Stickleback model}
\usage{
sb_fit(sb, sensors, events)
}
\arguments{
\item{sb}{\link{Stickleback}}

\item{sensors}{\link{Sensors}}

\item{events}{\link{Events}}
}
\description{
Fit a Stickleback model
}
