% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sb_plot_predictions}
\alias{sb_plot_predictions}
\title{Plot Stickleback predictions}
\usage{
sb_plot_predictions(deployid, sensors, predictions, outcomes = NULL)
}
\arguments{
\item{deployid}{\link{character(1)} Deployment ID.}

\item{sensors}{\link{Sensors} Sensor data.}

\item{predictions}{\link{Predictions}}

\item{outcomes}{\link{Outcomes}}
}
\value{

}
\description{
Plot Stickleback predictions
}
