% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sb_plot_predictions}
\alias{sb_plot_predictions}
\title{Plot Stickleback predictions}
\usage{
sb_plot_predictions(deployid, sensors, predictions, outcomes = NULL)
}
\arguments{
\item{deployid}{\verb{[character(1)]} Deployment ID.}

\item{sensors}{\verb{[Sensors]} Sensor data.}

\item{predictions}{\verb{[Predictions]}}

\item{outcomes}{\verb{[Outcomes]}}
}
\description{
Creates an interactive plotly figure for examining model predictions.
}
