% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stickleback.R
\name{sb_predict}
\alias{sb_predict}
\title{Make predictions with a Stickleback model}
\usage{
sb_predict(sb, sensors)
}
\arguments{
\item{sb}{\verb{[Stickleback]} A trained Stickleback model (see
\code{\link{Stickleback}}, \code{\link{sb_fit}})}

\item{sensors}{\verb{[Sensors]} Bio-logging sensor data (see
\code{\link{Sensors}})}
}
\value{
\verb{[Predictions]} The predicted events in the bio-logging sensor data
\code{sensors} from a trained Stickleback model \code{sb}.
}
\description{
Make predictions with a Stickleback model
}
\examples{
# Load sample data and split test/train
c(lunge_sensors, lunge_events) \%<-\% load_lunges()
test_deployids <- deployments(lunge_sensors)[1:3]
c(sensors_test, sensors_train) \%<-\% divide(lunge_sensors, test_deployids)
c(events_test, events_train) \%<-\% divide(lunge_events, test_deployids)

# Define a time series classifier
tsc <- compose_tsc(module = "interval_based",
                   algorithm = "SupervisedTimeSeriesForest",
                   params = list(n_estimators = 2L, random_state = 4321L),
                   columns = columns(lunge_sensors))

# Define a Stickleback model
sb <- Stickleback(tsc,
                  win_size = 50,
                  tol = 5,
                  nth = 10,
                  n_folds = 4,
                  seed = 1234)

# Fit the model to the sample data
sb_fit(sb, sensors_train, events_train)

# Use the model to make predictions
predictions <- sb_predict(sb, sensors_test)
predictions

}
\seealso{
\code{\link{sb_plot_predictions}}, \code{\link{sb_assess}}
}
