% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensors.R
\docType{class}
\name{Sensors-class}
\alias{Sensors-class}
\alias{Sensors}
\title{Sensors class}
\usage{
Sensors(sensor_data, deployid_col, datetime_col, sensor_cols)
}
\arguments{
\item{sensor_data}{\verb{[data.frame]} A data frame containing one or more
bio-logging sensor time series (e.g. pitch or overall dynamic body
acceleration). Must have columns for deployment ID and datetime (POSIXct).}

\item{deployid_col}{\verb{[character(1)]} Name of the column in \code{sensor_data} with
the bio-logger deployment IDs.}

\item{datetime_col}{\verb{[character(1)]} Name of the column in \code{sensor_data} with
datetimes of records.}

\item{sensor_cols}{\verb{[character]} Names of the columns in \code{sensor_data}
containing sensor data.}
}
\description{
Class representation for bio-logging sensor data.
}
\details{
Create a Sensors object with Sensors().
}
\section{Slots}{

\describe{
\item{\code{.data}}{A list of Pandas DataFrames.}
}}

\examples{
# Pretend we have bio-logging sensor data for two dogs
dogs_df <- data.frame(
  dogs = rep(c("A", "B"), each = 100),
  dogtime = rep(as.POSIXct("2000-01-01 06:00") + 1:100, 2),
  pitch = rep(90 * sin(seq(0, 20 * pi, length.out = 100))),
  odba = rep(90 * sin(seq(0, 20 * pi, length.out = 100)))^2
)
head(dogs_df)
Sensors(dogs_df, "dogs", "dogtime", c("pitch", "odba"))

}
