% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stickleback.R
\docType{class}
\name{Stickleback-class}
\alias{Stickleback-class}
\alias{Stickleback}
\title{Stickleback class}
\usage{
Stickleback(tsc, win_size, tol, nth = 1, n_folds = 4)
}
\arguments{
\item{tsc}{Python time series classifier. Must inherit sktime's BaseEstimator.}

\item{win_size}{integer. Sliding window size.}

\item{tol}{numeric. Prediction tolerance, in seconds.}

\item{nth}{integer. Sliding window step size.}

\item{n_folds}{integer. Number of folds for global cross validation step.}
}
\description{
R wrapper class for the Python implementation of the Stickleback method, used
for automated detection of behavioral events in bio-logging data.
}
\details{
Create a Stickleback object with Stickleback(). Fit it to labeled data with
fit(). Predict on new data with predict().
}
\section{Slots}{

\describe{
\item{\code{local_clf}}{Python time series classifier. Must inherit sktime's
BaseEstimator.}

\item{\code{win_size}}{integer. Sliding window size.}

\item{\code{tol}}{numeric. Prediction tolerance, in seconds.}

\item{\code{nth}}{integer. Sliding window step size.}

\item{\code{n_folds}}{integer. Number of folds for global cross validation step.}

\item{\code{.stickleback}}{Python Stickleback object.}
}}

